refactoring controller and routes

+++ app/controllers/job_applications_controller.rb
@@ -3,17 +3,16 @@ class JobApplicationsController < ApplicationController
   def create
     job = Job.find(params[:job_id])
     if application_exists(current_muso, job) || !current_muso
-      redirect_to "/jobs" and return if Job.find(params[:job_id])
+      redirect_to jobs_path and return if Job.find(params[:job_id])
     end
     job_app = JobApplication.new(
-      job: Job.find(params[:job_id]),
+      job: job,
       muso: current_muso,
       rejected: false 
     )
-   
     if job_app.save
       NotificationMailer.new_job_app_email(job_app).deliver_later
-      redirect_to '/jobs'
+      redirect_to jobs_path
     else
       @jobs = Job.all
       render '/jobs/index'
@@ -22,8 +21,9 @@ class JobApplicationsController < ApplicationController
 
   def destroy
     job_app = JobApplication.find(params[:job_application_id])
+    redirect_to jobs_path and return unless my_job_app?(job_app)
     if job_app.destroy
-      redirect_to '/jobs'
+      redirect_to jobs_path
     else
       @jobs = Job.all
       render '/jobs/index'
@@ -32,6 +32,7 @@ class JobApplicationsController < ApplicationController
 
   def reject
     job_app = JobApplication.find(params[:job_application_id])
+    redirect_to jobs_path and return unless my_job?(job_app.job)
     job_app.rejected = true
     if job_app.save
       redirect_to "/jobs/#{params[:id]}"
